backend/requirements.txt
fastapi[standard]
sqlalchemy
pydantic
polars
python-dotenv
psycopg2-binary